+++ libs/web/calendar/ui/quick-info-content/src/lib/quick-info-content-event/quick-info-content-event.component.ts
@@ -54,7 +54,7 @@ export class QuickInfoContentEventComponent implements OnInit {
     readonly schedule: ScheduleComponent,
     @Inject(Injector) private readonly injector: Injector,
     @Inject(TuiDialogService) private readonly dialogService: TuiDialogService,
-    private readonly alertService: TuiAlertService,
+    @Inject(TuiAlertService) private readonly alertService: TuiAlertService,
     private readonly store: QuickInfoContentEventStore,
     private readonly destroy$: TuiDestroyService
   ) {
@@ -90,6 +90,7 @@ export class QuickInfoContentEventComponent implements OnInit {
     if (this.data.Note !== this.initialEventNote) {
       this.store.updateNote({
         id: this.data.Id,
+        type: this.data.Type,
         payload: { note: this.data.Note },
       });
       this.schedule.saveEvent(this.data);