+++ libs/web/shared/data-access/models/src/lib/change-schedule/change-schedule-search.model.ts
@@ -1,6 +1,5 @@
 export type ChangeScheduleSearch = {
   status: number[];
   page: number;
-  pagination: number;
   teacherId?: string;
 };