+++ libs/web/teaching-schedule/ui/change-set-room-dialog/src/lib/change-set-room-dialog.module.ts
@@ -1,11 +1,13 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
+import { ReactiveComponentModule } from '@ngrx/component';
 import { TuiButtonModule } from '@taiga-ui/core';
 import { TuiComboBoxModule, TuiDataListWrapperModule } from '@taiga-ui/kit';
 import { FilterByInputPipeModule } from '@teaching-scheduling-system/core/pipes/filter-by-input-pipe';
 import { ChangeSetRoomDialogComponent } from './change-set-room-dialog.component';
 
+const NGRX = [ReactiveComponentModule];
 const TAIGA_UI = [TuiComboBoxModule, TuiDataListWrapperModule, TuiButtonModule];
 
 @NgModule({
@@ -13,7 +15,8 @@ const TAIGA_UI = [TuiComboBoxModule, TuiDataListWrapperModule, TuiButtonModule];
     CommonModule,
     ReactiveFormsModule,
     FilterByInputPipeModule,
-    TAIGA_UI,
+    ...NGRX,
+    ...TAIGA_UI,
   ],
   declarations: [ChangeSetRoomDialogComponent],
   exports: [ChangeSetRoomDialogComponent],